daisy: QCamera2: HAL3: Modify timeout logic
Issue:
As per the current timeout logic if the buffer
does not return to HAL in 3 sec it will report it as
missing frame. But for some cases such as long exposure
it takes almost 30 sec for buffer to return.

Fix:
Modifying the timeout logic to incorporate sensor
exposure time also.

Change-Id: Ic01c0b39ddaf64f2872a2b75ceffffa4d4ebc28b
Signed-off-by: TogoFire <togofire@mailfence.com>